refactor: complete migration from profiles to configs - update imports and remove profiles module
