+++ src/sboxmgr/cli/commands/export/cli.py
@@ -27,7 +27,7 @@ from .validators import (
 
 # Import Phase 3 components
 try:
-    from sboxmgr.profiles.models import FullProfile
+    from sboxmgr.configs.models import FullProfile
 
     PHASE3_AVAILABLE = True
 except ImportError:
