+++ src/sboxmgr/cli/commands/export/config_generators.py
@@ -18,7 +18,7 @@ from .validators import validate_middleware, validate_postprocessors
 
 # Import Phase 3 components
 try:
-    from sboxmgr.profiles.models import FullProfile
+    from sboxmgr.configs.models import FullProfile
 
     PHASE3_AVAILABLE = True
 except ImportError:
