+++ src/sboxmgr/cli/commands/export/profile_loaders.py
@@ -11,7 +11,7 @@ from sboxmgr.subscription.models import ClientProfile, InboundProfile
 
 # Import Phase 3 components
 try:
-    from sboxmgr.profiles.models import FullProfile
+    from sboxmgr.configs.models import FullProfile
 
     PHASE3_AVAILABLE = True
 except ImportError:
