--- src/sboxmgr/profiles/__init__.py
@@ -1,29 +1,0 @@
-"""Profile management module for sboxmgr (ADR-0017).
-
-This module provides profile-based configuration management,
-allowing users to define complete configurations in a single JSON file.
-"""
-
-from .models import (
-    AgentProfile,
-    ExportFormat,
-    ExportProfile,
-    FilterProfile,
-    FullProfile,
-    RoutingProfile,
-    SubscriptionProfile,
-    UIMode,
-    UIProfile,
-)
-
-__all__ = [
-    "FullProfile",
-    "SubscriptionProfile",
-    "FilterProfile",
-    "RoutingProfile",
-    "ExportProfile",
-    "AgentProfile",
-    "UIProfile",
-    "ExportFormat",
-    "UIMode",
-]
